refactor(apis): api 함수 분리
