--- src/apis/api.js
@@ -1,29 +1,0 @@
-import axios from "axios";
-
-const BASE_URL = `https://bootcamp-api.codeit.kr`;
-const USER_ID = `users/1`;
-
-export async function getUser() {
-  const response = await axios.get(`${BASE_URL}/api/${USER_ID}`);
-  // const body = await response.json();
-  return response.data;
-}
-
-export async function getAllFolders() {
-  const response = await axios.get(`${BASE_URL}/api/${USER_ID}/folders`);
-  // const body = await response.json();
-  return response.data;
-}
-
-export async function getAllCards(id = "") {
-  let url = "";
-  if (id === "") {
-    url = `${BASE_URL}/api/${USER_ID}/links`;
-  } else {
-    const query = `?folderId=${id}`;
-    url = `${BASE_URL}/api/${USER_ID}/links${query}`;
-  }
-  const response = await axios.get(`${url}`);
-  // const body = await response.json();
-  return response.data;
-}
